refactor: Implement useLogin hook
