+++ client/src/Pages/Login/Login.js
@@ -1,73 +1,26 @@
-import React, { useContext, useState } from 'react'
-
-import { login } from '_Constants'
-
-import {  PasswordInput, Input, Loading } from 'Components'
+import {  PasswordInput, Input } from 'Components'
 import { AuthCard } from 'Components'
 
 import { AiOutlineLogin } from 'react-icons/ai'
 import { Link, useNavigate, useLocation } from 'react-router-dom'
 
-import AuthContext from '../../Context/AuthProvider'
-
-// tacc/Abcd@1234
+import { useLogin } from 'Hooks'
 
 export default function Login() {
-  const [userName, setUserName] = useState('')
-  const [password, setPassword] = useState('')
-  const [error, setError] = useState([])
-
-  const [isDisabled, setIsDisabled] = useState(false)
-
-  const { setUser } = useContext(AuthContext)
-  const navigate = useNavigate()
   const location = useLocation()
   const from = location.state?.from.pathname || '/'
+  const navigate = useNavigate()
 
-  const handleLogin = (e) =>{
-    e.preventDefault()
-    const userCredentials = {userName,password}
-    setIsDisabled(true)
-    fetch(`${login}`, {
-      method: 'POST',
-      headers: {
-        "Content-Type": "application/json",
-      },
-      credentials: 'include',
-      body: JSON.stringify(userCredentials)
-    })
-    .then(response=>{
-      if (response.ok) {
-        return response.json()
-      } else {
-        throw response
-      }
-    })
-    .then(response=>{
-      const {roles, userName } = response
-
-      const Identity = {
-        userName,
-        roles,
-        unAuthorized: false
-      }
-
-      setIsDisabled(false)
-      setUser(Identity)
-      navigate(from, {replace: true})
-    })
-    .catch(error=>{
-      setIsDisabled(false)
-      if (error instanceof Response) {
-        error.json().then(errorData => {
-          //const errorMessages = errorData.map(e=>e.description)
-          setError(errorData.description)
-        })
-      } else {
-        console.error('Error:', error)
-      }
-    })
-  }
+  const {
+    userName,
+    setUserName,
+    password,
+    setPassword,
+    error,
+    setError,
+    isDisabled,
+    handleLogin
+  } = useLogin(navigate, from)
 
   return (
     <>
